@@ -1,6 +1,6 @@
 {
   "name": "prettier",
-  "version": "1.13.7",
+  "version": "1.14.0",
   "description": "Prettier is an opinionated code formatter",
   "bin": {
     "prettier": "./bin/prettier.js"
@@ -14,7 +14,7 @@
     "node": ">=6"
   },
   "dependencies": {
-    "@babel/code-frame": "7.0.0-beta.49",
+    "@babel/code-frame": "7.0.0-beta.46",
     "@babel/parser": "7.0.0-beta.49",
     "@glimmer/syntax": "0.30.3",
     "camelcase": "4.1.0",
@@ -38,32 +38,38 @@
     "graphql": "0.13.2",
     "html-tag-names": "1.1.2",
     "ignore": "3.3.7",
-    "jest-docblock": "22.2.2",
+    "jest-docblock": "23.2.0",
     "json-stable-stringify": "1.0.1",
     "leven": "2.1.0",
+    "linguist-languages": "6.2.1-dev.20180706",
     "lodash.uniqby": "4.7.0",
     "mem": "1.1.0",
     "minimatch": "3.0.4",
     "minimist": "1.2.0",
+    "normalize-path": "3.0.0",
     "parse5": "3.0.3",
     "postcss-less": "1.1.5",
     "postcss-media-query-parser": "0.2.3",
-    "postcss-scss": "1.0.5",
+    "postcss-scss": "1.0.6",
     "postcss-selector-parser": "2.2.3",
     "postcss-values-parser": "1.5.0",
     "remark-parse": "5.0.0",
     "resolve": "1.5.0",
     "semver": "5.4.1",
     "string-width": "2.1.1",
-    "typescript": "2.9.0-dev.20180421",
-    "typescript-eslint-parser": "16.0.0",
+    "typescript": "3.0.0-dev.20180626",
+    "typescript-eslint-parser": "17.0.0",
     "unicode-regex": "1.0.1",
-    "unified": "6.1.6"
+    "unified": "6.1.6",
+    "yaml": "1.0.0-rc.7",
+    "yaml-unist-parser": "1.0.0-rc.2"
   },
   "devDependencies": {
     "@babel/cli": "7.0.0-beta.49",
     "@babel/core": "7.0.0-beta.49",
     "@babel/preset-env": "7.0.0-beta.49",
+    "babel-loader": "8.0.0-beta.3",
+    "benchmark": "2.1.4",
     "builtin-modules": "2.0.0",
     "codecov": "2.2.0",
     "cross-env": "5.0.5",
@@ -73,13 +79,17 @@
     "eslint-plugin-import": "2.9.0",
     "eslint-plugin-prettier": "2.6.0",
     "eslint-plugin-react": "7.7.0",
+    "execa": "0.10.0",
     "fs2": "^0.2.7",
-    "jest": "21.1.0",
+    "jest": "23.3.0",
+    "jest-junit": "5.0.0",
+    "jest-watch-typeahead": "0.1.0",
     "mkdirp": "0.5.1",
-    "prettier": "1.13.4",
+    "prettier": "1.13.7",
     "prettylint": "1.0.0",
     "rimraf": "2.6.2",
     "rollup": "0.47.6",
+    "rollup-plugin-alias": "1.4.0",
     "rollup-plugin-babel": "4.0.0-beta.4",
     "rollup-plugin-commonjs": "8.2.6",
     "rollup-plugin-json": "2.1.1",
@@ -89,10 +99,13 @@
     "rollup-plugin-replace": "1.2.1",
     "rollup-plugin-uglify": "3.0.0",
     "shelljs": "0.8.1",
-    "snapshot-diff": "0.2.2",
+    "snapshot-diff": "0.4.0",
     "strip-ansi": "4.0.0",
     "tempy": "0.2.1",
-    "webpack": "2.6.1"
+    "webpack": "3.12.0"
+  },
+  "resolutions": {
+    "@babel/code-frame": "7.0.0-beta.46"
   },
   "scripts": {
     "prepublishOnly": "echo \"Error: must publish from dist/\" && exit 1",
@@ -100,6 +113,9 @@
     "test": "jest",
     "test:dist": "node ./scripts/test-dist.js",
     "test-integration": "jest tests_integration",
+    "perf-repeat": "yarn && yarn build && cross-env NODE_ENV=production node ./dist/bin-prettier.js --debug-repeat ${PERF_REPEAT:-1000} --loglevel debug ${PERF_FILE:-./index.js} > /dev/null",
+    "perf-repeat-inspect": "yarn && yarn build && cross-env NODE_ENV=production node --inspect-brk ./dist/bin-prettier.js --debug-repeat ${PERF_REPEAT:-1000} --loglevel debug ${PERF_FILE:-./index.js} > /dev/null",
+    "perf-benchmark": "yarn && yarn build && cross-env NODE_ENV=production node ./dist/bin-prettier.js --debug-benchmark --loglevel debug ${PERF_FILE:-./index.js} > /dev/null",
     "lint": "cross-env EFF_NO_LINK_RULES=true eslint . --format node_modules/eslint-friendly-formatter",
     "lint-docs": "prettylint {.,docs,website,website/blog}/*.md",
     "build": "node ./scripts/build/build.js",
