@@ -1,6 +1,7 @@
 "use strict";
 
 const React = require("react");
+const PropTypes = require("prop-types");
 
 const CompLibrary = require("../../core/CompLibrary.js");
 const MarkdownBlock = CompLibrary.MarkdownBlock;
@@ -14,7 +15,7 @@ const ButtonGroup = props => (
 );
 
 ButtonGroup.propTypes = {
-  children: React.PropTypes.node
+  children: PropTypes.node
 };
 
 class Button extends React.Component {
@@ -34,9 +35,9 @@ Button.defaultProps = {
 };
 
 Button.propTypes = {
-  href: React.PropTypes.string,
-  target: React.PropTypes.string,
-  children: React.PropTypes.any
+  href: PropTypes.string,
+  target: PropTypes.string,
+  children: PropTypes.any
 };
 
 const HomeSplash = props => {
@@ -69,7 +70,7 @@ const HomeSplash = props => {
 };
 
 HomeSplash.propTypes = {
-  language: React.PropTypes.string
+  language: PropTypes.string
 };
 
 const TldrSection = ({ language }) => (
@@ -108,7 +109,7 @@ const TldrSection = ({ language }) => (
 );
 
 TldrSection.propTypes = {
-  language: React.PropTypes.string
+  language: PropTypes.string
 };
 
 const Language = ({ name, showName, image, variants }) => (
@@ -124,19 +125,19 @@ const Language = ({ name, showName, image, variants }) => (
     <div>
       {showName && <p className="accented">{name}</p>}
       {variants.map(variant => (
-        <code key={variant}>
+        <div key={variant}>
           <MarkdownBlock>{variant}</MarkdownBlock>
-        </code>
+        </div>
       ))}
     </div>
   </div>
 );
 
 Language.propTypes = {
-  name: React.PropTypes.string,
-  showName: React.PropTypes.boolean,
-  image: React.PropTypes.string,
-  variants: React.PropTypes.array
+  name: PropTypes.string,
+  showName: PropTypes.boolean,
+  image: PropTypes.string,
+  variants: PropTypes.array
 };
 
 const LanguagesSection = () => {
@@ -190,16 +191,16 @@ const Editor = ({ content = "", image, name }) => (
   <div className="editor">
     <img className="editorImage" src={image} />
     <div className="editorInfo">
-      <h3 style={{ marginBottom: "-16px" }}>{name}</h3>
+      <h3 className="editorName">{name}</h3>
       <MarkdownBlock>{content.replace(/\n/g, "  \n")}</MarkdownBlock>
     </div>
   </div>
 );
 
 Editor.propTypes = {
-  content: React.PropTypes.string,
-  image: React.PropTypes.string.isRequired,
-  name: React.PropTypes.string.isRequired
+  content: PropTypes.string,
+  image: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired
 };
 
 const EditorSupportSection = () => (
@@ -428,7 +429,7 @@ const UsersSection = ({ language }) => {
 };
 
 UsersSection.propTypes = {
-  language: React.PropTypes.string
+  language: PropTypes.string
 };
 
 class Index extends React.Component {
@@ -439,7 +440,7 @@ class Index extends React.Component {
       <div>
         <script src="landing.js" />
         <HomeSplash language={language} />
-        <div className="mainContainer">
+        <div className="mainContainer landingContainer">
           <TldrSection language={language} />
           <LanguagesSection />
           <EditorSupportSection />
@@ -452,7 +453,7 @@ class Index extends React.Component {
 }
 
 Index.propTypes = {
-  language: React.PropTypes.string
+  language: PropTypes.string
 };
 
 module.exports = Index;
