@@ -4,7 +4,7 @@ const chalk = require("chalk");
 const { runYarn, logPromise, readJson } = require("../utils");
 
 module.exports = async function({ version }) {
-  await logPromise("Generating bundles", runYarn("build"));
+  await logPromise("Generating bundles", runYarn(["build", "--purge-cache"]));
 
   const builtPkg = await readJson("dist/package.json");
   if (builtPkg.version !== version) {
