@@ -1,20 +1,19 @@
 "use strict";
 
 const execa = require("execa");
-const { readFileSync, writeFileSync } = require("fs");
-const { logPromise, readJson, writeJson } = require("../utils");
+const { logPromise, readJson, writeJson, processFile } = require("../utils");
 
 async function bump({ version }) {
   const pkg = await readJson("package.json");
   pkg.version = version;
   await writeJson("package.json", pkg, { spaces: 2 });
 
-  // Update .github/ISSUE_TEMPLATE.MD
-  const issueFile = ".github/ISSUE_TEMPLATE.md";
-  const issueTempl = readFileSync(issueFile, "utf-8");
-  writeFileSync(
-    issueFile,
-    issueTempl.replace(/^\*\*Prettier .*?\*\*$/m, `**Prettier ${version}**`)
+  // Update github issue templates
+  processFile(".github/ISSUE_TEMPLATE/formatting.md", content =>
+    content.replace(/^(\*\*Prettier ).*?(\*\*)$/m, `$1${version}$2`)
+  );
+  processFile(".github/ISSUE_TEMPLATE/integration.md", content =>
+    content.replace(/^(- Prettier Version: ).*?$/m, `$1${version}`)
   );
 }
 
