@@ -20,7 +20,11 @@ shell.exec("npm init -y", { cwd: tmpDir });
 shell.exec(`npm install "${tarPath}"`, { cwd: tmpDir });
 shell.config.silent = false;
 
-const code = shell.exec("yarn test --color --runInBand", {
+const cmd = `yarn test --color --runInBand ${
+  process.env.TEST_STANDALONE ? "tests/" : ""
+}`;
+
+const code = shell.exec(cmd, {
   cwd: rootDir,
   env: Object.assign({}, process.env, {
     NODE_ENV: "production",
