@@ -1,9 +1,12 @@
 "use strict";
 
-const util = require("../common/util");
+const { getStringWidth } = require("../common/util");
 const { concat, fill, cursor } = require("./doc-builders");
 
 const customizations = require("../_customizations/doc-printer");
+
+/** @type {{[groupId: PropertyKey]: MODE}} */
+let groupModeMap;
 
 const MODE_BREAK = 1;
 const MODE_FLAT = 2;
@@ -129,7 +132,7 @@ function fits(next, restCommands, width, options, mustBeFlat) {
     const doc = x[2];
 
     if (typeof doc === "string") {
-      width -= util.getStringWidth(doc);
+      width -= getStringWidth(doc);
     } else {
       switch (doc.type) {
         case "concat":
@@ -152,6 +155,9 @@ function fits(next, restCommands, width, options, mustBeFlat) {
           }
           cmds.push([ind, doc.break ? MODE_BREAK : mode, doc.contents]);
 
+          if (doc.id) {
+            groupModeMap[doc.id] = cmds[cmds.length - 1][1];
+          }
           break;
         case "fill":
           for (let i = doc.parts.length - 1; i >= 0; i--) {
@@ -159,19 +165,21 @@ function fits(next, restCommands, width, options, mustBeFlat) {
           }
 
           break;
-        case "if-break":
-          if (mode === MODE_BREAK) {
+        case "if-break": {
+          const groupMode = doc.groupId ? groupModeMap[doc.groupId] : mode;
+          if (groupMode === MODE_BREAK) {
             if (doc.breakContents) {
               cmds.push([ind, mode, doc.breakContents]);
             }
           }
-          if (mode === MODE_FLAT) {
+          if (groupMode === MODE_FLAT) {
             if (doc.flatContents) {
               cmds.push([ind, mode, doc.flatContents]);
             }
           }
 
           break;
+        }
         case "line":
           switch (mode) {
             // fallthrough
@@ -196,6 +204,8 @@ function fits(next, restCommands, width, options, mustBeFlat) {
 }
 
 function printDocToString(doc, options) {
+  groupModeMap = {};
+
   const width = options.printWidth;
   const newLine = options.newLine || "\n";
   let pos = 0;
@@ -216,7 +226,7 @@ function printDocToString(doc, options) {
     if (typeof doc === "string") {
       out.push(doc);
 
-      pos += util.getStringWidth(doc);
+      pos += getStringWidth(doc);
     } else {
       switch (doc.type) {
         case "cursor":
@@ -318,6 +328,10 @@ function printDocToString(doc, options) {
               break;
             }
           }
+
+          if (doc.id) {
+            groupModeMap[doc.id] = cmds[cmds.length - 1][1];
+          }
           break;
         // Fills each line with as much code as possible before moving to a new
         // line with the same indentation.
@@ -414,19 +428,21 @@ function printDocToString(doc, options) {
           }
           break;
         }
-        case "if-break":
-          if (mode === MODE_BREAK) {
+        case "if-break": {
+          const groupMode = doc.groupId ? groupModeMap[doc.groupId] : mode;
+          if (groupMode === MODE_BREAK) {
             if (doc.breakContents) {
               cmds.push([ind, mode, doc.breakContents]);
             }
           }
-          if (mode === MODE_FLAT) {
+          if (groupMode === MODE_FLAT) {
             if (doc.flatContents) {
               cmds.push([ind, mode, doc.flatContents]);
             }
           }
 
           break;
+        }
         case "line-suffix":
           lineSuffix.push([ind, mode, doc.contents]);
           break;
