@@ -1,7 +1,19 @@
 "use strict";
 
+// Using a unique object to compare by reference.
+const traverseDocOnExitStackMarker = {};
+
 function traverseDoc(doc, onEnter, onExit, shouldTraverseConditionalGroups) {
-  function traverseDocRec(doc) {
+  const docsStack = [doc];
+
+  while (docsStack.length !== 0) {
+    const doc = docsStack.pop();
+
+    if (doc === traverseDocOnExitStackMarker) {
+      onExit(docsStack.pop());
+      continue;
+    }
+
     let shouldRecurse = true;
     if (onEnter) {
       if (onEnter(doc) === false) {
@@ -9,35 +21,40 @@ function traverseDoc(doc, onEnter, onExit, shouldTraverseConditionalGroups) {
       }
     }
 
+    if (onExit) {
+      docsStack.push(doc);
+      docsStack.push(traverseDocOnExitStackMarker);
+    }
+
     if (shouldRecurse) {
+      // When there are multiple parts to process,
+      // the parts need to be pushed onto the stack in reverse order,
+      // so that they are processed in the original order
+      // when the stack is popped.
       if (doc.type === "concat" || doc.type === "fill") {
-        for (let i = 0; i < doc.parts.length; i++) {
-          traverseDocRec(doc.parts[i]);
+        for (let ic = doc.parts.length, i = ic - 1; i >= 0; --i) {
+          docsStack.push(doc.parts[i]);
         }
       } else if (doc.type === "if-break") {
-        if (doc.breakContents) {
-          traverseDocRec(doc.breakContents);
-        }
         if (doc.flatContents) {
-          traverseDocRec(doc.flatContents);
+          docsStack.push(doc.flatContents);
+        }
+        if (doc.breakContents) {
+          docsStack.push(doc.breakContents);
         }
       } else if (doc.type === "group" && doc.expandedStates) {
         if (shouldTraverseConditionalGroups) {
-          doc.expandedStates.forEach(traverseDocRec);
+          for (let ic = doc.expandedStates.length, i = ic - 1; i >= 0; --i) {
+            docsStack.push(doc.expandedStates[i]);
+          }
         } else {
-          traverseDocRec(doc.contents);
+          docsStack.push(doc.contents);
         }
       } else if (doc.contents) {
-        traverseDocRec(doc.contents);
+        docsStack.push(doc.contents);
       }
     }
-
-    if (onExit) {
-      onExit(doc);
-    }
   }
-
-  traverseDocRec(doc);
 }
 
 function mapDoc(doc, cb) {
@@ -58,7 +75,7 @@ function mapDoc(doc, cb) {
 function findInDoc(doc, fn, defaultValue) {
   let result = defaultValue;
   let hasStopped = false;
-  traverseDoc(doc, doc => {
+  function findInDocOnEnterFn(doc) {
     const maybeResult = fn(doc);
     if (maybeResult !== undefined) {
       hasStopped = true;
@@ -67,7 +84,8 @@ function findInDoc(doc, fn, defaultValue) {
     if (hasStopped) {
       return false;
     }
-  });
+  }
+  traverseDoc(doc, findInDocOnEnterFn);
   return result;
 }
 
@@ -75,37 +93,33 @@ function isEmpty(n) {
   return typeof n === "string" && n.length === 0;
 }
 
+function isLineNextFn(doc) {
+  if (typeof doc === "string") {
+    return false;
+  }
+  if (doc.type === "line") {
+    return true;
+  }
+}
+
 function isLineNext(doc) {
-  return findInDoc(
-    doc,
-    doc => {
-      if (typeof doc === "string") {
-        return false;
-      }
-      if (doc.type === "line") {
-        return true;
-      }
-    },
-    false
-  );
+  return findInDoc(doc, isLineNextFn, false);
+}
+
+function willBreakFn(doc) {
+  if (doc.type === "group" && doc.break) {
+    return true;
+  }
+  if (doc.type === "line" && doc.hard) {
+    return true;
+  }
+  if (doc.type === "break-parent") {
+    return true;
+  }
 }
 
 function willBreak(doc) {
-  return findInDoc(
-    doc,
-    doc => {
-      if (doc.type === "group" && doc.break) {
-        return true;
-      }
-      if (doc.type === "line" && doc.hard) {
-        return true;
-      }
-      if (doc.type === "break-parent") {
-        return true;
-      }
-    },
-    false
-  );
+  return findInDoc(doc, willBreakFn, false);
 }
 
 function breakParentGroup(groupStack) {
@@ -121,47 +135,51 @@ function breakParentGroup(groupStack) {
 }
 
 function propagateBreaks(doc) {
-  const alreadyVisited = new Map();
+  const alreadyVisitedSet = new Set();
   const groupStack = [];
-  traverseDoc(
-    doc,
-    doc => {
-      if (doc.type === "break-parent") {
+  function propagateBreaksOnEnterFn(doc) {
+    if (doc.type === "break-parent") {
+      breakParentGroup(groupStack);
+    }
+    if (doc.type === "group") {
+      groupStack.push(doc);
+      if (alreadyVisitedSet.has(doc)) {
+        return false;
+      }
+      alreadyVisitedSet.add(doc);
+    }
+  }
+  function propagateBreaksOnExitFn(doc) {
+    if (doc.type === "group") {
+      const group = groupStack.pop();
+      if (group.break) {
         breakParentGroup(groupStack);
       }
-      if (doc.type === "group") {
-        groupStack.push(doc);
-        if (alreadyVisited.has(doc)) {
-          return false;
-        }
-        alreadyVisited.set(doc, true);
-      }
-    },
-    doc => {
-      if (doc.type === "group") {
-        const group = groupStack.pop();
-        if (group.break) {
-          breakParentGroup(groupStack);
-        }
-      }
-    },
+    }
+  }
+  traverseDoc(
+    doc,
+    propagateBreaksOnEnterFn,
+    propagateBreaksOnExitFn,
     /* shouldTraverseConditionalGroups */ true
   );
 }
 
-function removeLines(doc) {
+function removeLinesFn(doc) {
   // Force this doc into flat mode by statically converting all
   // lines into spaces (or soft lines into nothing). Hard lines
   // should still output because there's too great of a chance
   // of breaking existing assumptions otherwise.
-  return mapDoc(doc, d => {
-    if (d.type === "line" && !d.hard) {
-      return d.soft ? "" : " ";
-    } else if (d.type === "if-break") {
-      return d.flatContents || "";
-    }
-    return d;
-  });
+  if (doc.type === "line" && !doc.hard) {
+    return doc.soft ? "" : " ";
+  } else if (doc.type === "if-break") {
+    return doc.flatContents || "";
+  }
+  return doc;
+}
+
+function removeLines(doc) {
+  return mapDoc(doc, removeLinesFn);
 }
 
 function stripTrailingHardline(doc) {
