@@ -2,50 +2,30 @@
 
 const printer = require("./printer-postcss");
 const options = require("./options");
-
-// Based on:
-// https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
+const languageExtend = require("../utils/language-extend");
 
 const languages = [
-  {
-    name: "CSS",
+  languageExtend({}, require("linguist-languages/data/css"), {
     since: "1.4.0",
     parsers: ["css"],
-    group: "CSS",
-    tmScope: "source.css",
-    aceMode: "css",
-    codemirrorMode: "css",
-    codemirrorMimeType: "text/css",
-    extensions: [".css", ".pcss", ".postcss"],
-    liguistLanguageId: 50,
-    vscodeLanguageIds: ["css", "postcss"]
-  },
-  {
-    name: "Less",
+    vscodeLanguageIds: ["css"]
+  }),
+  languageExtend({}, require("linguist-languages/data/postcss"), {
+    since: "1.4.0",
+    parsers: ["css"],
+    extensions: [".postcss"],
+    vscodeLanguageIds: ["postcss"]
+  }),
+  languageExtend({}, require("linguist-languages/data/less"), {
     since: "1.4.0",
     parsers: ["less"],
-    group: "CSS",
-    extensions: [".less"],
-    tmScope: "source.css.less",
-    aceMode: "less",
-    codemirrorMode: "css",
-    codemirrorMimeType: "text/css",
-    liguistLanguageId: 198,
     vscodeLanguageIds: ["less"]
-  },
-  {
-    name: "SCSS",
+  }),
+  languageExtend({}, require("linguist-languages/data/scss"), {
     since: "1.4.0",
     parsers: ["scss"],
-    group: "CSS",
-    tmScope: "source.scss",
-    aceMode: "scss",
-    codemirrorMode: "css",
-    codemirrorMimeType: "text/x-scss",
-    extensions: [".scss"],
-    liguistLanguageId: 329,
     vscodeLanguageIds: ["scss"]
-  }
+  })
 ];
 
 const printers = {
