@@ -3,6 +3,7 @@
 const createError = require("../common/parser-create-error");
 const parseFrontMatter = require("../utils/front-matter");
 const lineColumnToIndex = require("../utils/line-column-to-index");
+const { hasPragma } = require("./pragma");
 
 // utils
 const utils = require("./utils");
@@ -346,7 +347,7 @@ function parseNestedCSS(node) {
       return node;
     }
 
-    if (node.type !== "css-comment-yaml" && value.length > 0) {
+    if (value.length > 0) {
       const defaultSCSSDirectiveIndex = value.match(DEFAULT_SCSS_DIRECTIVE);
 
       if (defaultSCSSDirectiveIndex) {
@@ -479,7 +480,7 @@ function parseNestedCSS(node) {
 
 function parseWithParser(parser, text) {
   const parsed = parseFrontMatter(text);
-  const frontMatter = parsed.frontMatter;
+  const { frontMatter } = parsed;
   text = parsed.content;
 
   let result;
@@ -496,10 +497,7 @@ function parseWithParser(parser, text) {
   result = parseNestedCSS(addTypePrefix(result, "css-"));
 
   if (frontMatter) {
-    result.nodes.unshift({
-      type: "front-matter",
-      value: frontMatter
-    });
+    result.nodes.unshift(frontMatter);
   }
 
   return result;
@@ -546,6 +544,7 @@ function parse(text, parsers, opts) {
 const parser = {
   parse,
   astFormat: "postcss",
+  hasPragma,
   locStart(node) {
     if (node.source) {
       return lineColumnToIndex(node.source.start, node.source.input.css) - 1;
