@@ -1,6 +1,8 @@
 "use strict";
 
 const clean = require("./clean");
+const embed = require("./embed");
+const { insertPragma } = require("./pragma");
 const {
   printNumber,
   printString,
@@ -92,8 +94,9 @@ function genericPrint(path, options, print) {
   }
 
   switch (node.type) {
-    case "front-matter":
-      return concat([node.value, hardline]);
+    case "yaml":
+    case "toml":
+      return concat([node.raw, hardline]);
     case "css-root": {
       const nodes = printNodeSequence(path, options, print);
 
@@ -376,7 +379,7 @@ function genericPrint(path, options, print) {
         node.namespace
           ? concat([node.namespace === true ? "" : node.namespace.trim(), "|"])
           : "",
-        adjustNumbers(node.value)
+        node.value
       ]);
     }
     case "selector-pseudo": {
@@ -832,7 +835,9 @@ function printNodeSequence(path, options, print) {
             options.originalText,
             options.locStart(node.nodes[i + 1]),
             { backwards: true }
-          )) ||
+          ) &&
+          node.nodes[i].type !== "yaml" &&
+          node.nodes[i].type !== "toml") ||
         (node.nodes[i + 1].type === "css-atrule" &&
           node.nodes[i + 1].name === "else" &&
           node.nodes[i].type !== "css-comment")
@@ -841,7 +846,13 @@ function printNodeSequence(path, options, print) {
       } else {
         parts.push(hardline);
         if (
-          isNextLineEmpty(options.originalText, pathChild.getValue(), options)
+          isNextLineEmpty(
+            options.originalText,
+            pathChild.getValue(),
+            options
+          ) &&
+          node.nodes[i].type !== "yaml" &&
+          node.nodes[i].type !== "toml"
         ) {
           parts.push(hardline);
         }
@@ -899,6 +910,8 @@ function printCssNumber(rawNumber) {
 
 module.exports = {
   print: genericPrint,
+  embed,
+  insertPragma,
   hasPrettierIgnore: hasIgnoreComment,
   massageAstNode: clean
 };
