@@ -1,21 +1,14 @@
 "use strict";
 
 const printer = require("./printer-glimmer");
-
-// Based on:
-// https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
+const languageExtend = require("../utils/language-extend");
 
 const languages = [
-  {
-    type: "markup",
-    group: "HTML",
-    aliases: ["hbs", "htmlbars"],
-    extensions: [".handlebars", ".hbs"],
-    tm_scope: "text.html.handlebars",
-    ace_mode: "handlebars",
-    language_id: 155,
-    since: null // unreleased
-  }
+  languageExtend({}, require("linguist-languages/data/handlebars"), {
+    since: null, // unreleased
+    parsers: ["glimmer"],
+    vscodeLanguageIds: ["handlebars"]
+  })
 ];
 
 const printers = {
