@@ -10,8 +10,10 @@ function parse(text, parsers, opts) {
 
   const babylonOptions = {
     sourceType: "module",
+    allowAwaitOutsideFunction: true,
     allowImportExportEverywhere: true,
     allowReturnOutsideFunction: true,
+    allowSuperOutsideMethod: true,
     plugins: [
       "jsx",
       "flow",
@@ -31,7 +33,9 @@ function parse(text, parsers, opts) {
       "optionalChaining",
       "classPrivateProperties",
       "pipelineOperator",
-      "nullishCoalescingOperator"
+      "nullishCoalescingOperator",
+      "bigInt",
+      "throwExpressions"
     ]
   };
 
