@@ -12,7 +12,8 @@ function embed(path, print, textToDoc, options) {
 
   if (node.type === "code" && node.lang !== null) {
     // only look for the first string so as to support [markdown-preview-enhanced](https://shd101wyy.github.io/markdown-preview-enhanced/#/code-chunk)
-    const lang = node.lang.split(/\s/, 1)[0];
+    const langMatch = node.lang.match(/^[A-Za-z0-9_-]+/);
+    const lang = langMatch ? langMatch[0] : "";
     const parser = getParserName(lang);
     if (parser) {
       const styleUnit = options.__inJsTemplate ? "~" : "`";
@@ -32,6 +33,21 @@ function embed(path, print, textToDoc, options) {
     }
   }
 
+  if (node.type === "yaml") {
+    return markAsRoot(
+      concat([
+        "---",
+        hardline,
+        node.value.trim()
+          ? replaceNewlinesWithLiterallines(
+              textToDoc(node.value, { parser: "yaml" })
+            )
+          : "",
+        "---"
+      ])
+    );
+  }
+
   return null;
 
   function getParserName(lang) {
@@ -41,6 +57,7 @@ function embed(path, print, textToDoc, options) {
     const language = supportInfo.languages.find(
       language =>
         language.name.toLowerCase() === lang ||
+        (language.aliases && language.aliases.indexOf(lang) !== -1) ||
         (language.extensions &&
           language.extensions.find(ext => ext.substring(1) === lang))
     );
