@@ -2,36 +2,15 @@
 
 const printer = require("./printer-markdown");
 const options = require("./options");
-
-// Based on:
-// https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
+const languageExtend = require("../utils/language-extend");
 
 const languages = [
-  {
-    name: "Markdown",
+  languageExtend({}, require("linguist-languages/data/markdown"), {
     since: "1.8.0",
     parsers: ["remark"],
-    aliases: ["pandoc"],
-    aceMode: "markdown",
-    codemirrorMode: "gfm",
-    codemirrorMimeType: "text/x-gfm",
-    wrap: true,
-    extensions: [
-      ".md",
-      ".markdown",
-      ".mdown",
-      ".mdwn",
-      ".mkd",
-      ".mkdn",
-      ".mkdown",
-      ".ron",
-      ".workbook"
-    ],
     filenames: ["README"],
-    tmScope: "source.gfm",
-    linguistLanguageId: 222,
     vscodeLanguageIds: ["markdown"]
-  }
+  })
 ];
 
 const printers = {
