@@ -5,6 +5,10 @@ const unified = require("unified");
 const pragma = require("./pragma");
 const parseFrontMatter = require("../utils/front-matter");
 const util = require("../common/util");
+const { getOrderedListItemInfo } = require("./utils");
+
+// 0x0 ~ 0x10ffff
+const isSingleCharRegex = /^([\u0000-\uffff]|[\ud800-\udbff][\udc00-\udfff])$/;
 
 /**
  * based on [MDAST](https://github.com/syntax-tree/mdast) with following modifications:
@@ -28,18 +32,23 @@ function parse(text, parsers, opts) {
     .use(restoreUnescapedCharacter(text))
     .use(mergeContinuousTexts)
     .use(transformInlineCode)
+    .use(transformIndentedCodeblockAndMarkItsParentList(text))
+    .use(markAlignedList(text, opts))
     .use(splitText(opts));
   return processor.runSync(processor.parse(text));
 }
 
 function map(ast, handler) {
-  return (function preorder(node, index, parentNode) {
-    const newNode = Object.assign({}, handler(node, index, parentNode));
+  return (function preorder(node, index, parentStack) {
+    parentStack = parentStack || [];
+
+    const newNode = Object.assign({}, handler(node, index, parentStack));
     if (newNode.children) {
       newNode.children = newNode.children.map((child, index) => {
-        return preorder(child, index, newNode);
+        return preorder(child, index, [newNode].concat(parentStack));
       });
     }
+
     return newNode;
   })(ast, null, null);
 }
@@ -66,8 +75,9 @@ function restoreUnescapedCharacter(originalText) {
             value:
               node.value !== "*" &&
               node.value !== "_" && // handle these two cases in printer
-              node.value.length === 1 &&
-              node.position.end.offset - node.position.start.offset > 1
+              isSingleCharRegex.test(node.value) &&
+              node.position.end.offset - node.position.start.offset !==
+                node.value.length
                 ? originalText.slice(
                     node.position.start.offset,
                     node.position.end.offset
@@ -105,7 +115,7 @@ function mergeContinuousTexts() {
 
 function splitText(options) {
   return () => ast =>
-    map(ast, (node, index, parentNode) => {
+    map(ast, (node, index, [parentNode]) => {
       if (node.type !== "text") {
         return node;
       }
@@ -138,10 +148,7 @@ function frontMatter() {
     const parsed = parseFrontMatter(value);
 
     if (parsed.frontMatter) {
-      return eat(parsed.frontMatter)({
-        type: "front-matter",
-        value: parsed.frontMatter
-      });
+      return eat(parsed.frontMatter.raw)(parsed.frontMatter);
     }
   }
   tokenizer.onlyAtStart = true;
@@ -166,6 +173,149 @@ function liquid() {
   tokenizer.locator = function(value, fromIndex) {
     return value.indexOf("{", fromIndex);
   };
+}
+
+function transformIndentedCodeblockAndMarkItsParentList(originalText) {
+  return () => ast =>
+    map(ast, (node, index, parentStack) => {
+      if (node.type === "code") {
+        // the first char may point to `\n`, e.g. `\n\t\tbar`, just ignore it
+        const isIndented = /^\n?( {4,}|\t)/.test(
+          originalText.slice(
+            node.position.start.offset,
+            node.position.end.offset
+          )
+        );
+
+        node.isIndented = isIndented;
+
+        if (isIndented) {
+          for (let i = 0; i < parentStack.length; i++) {
+            const parent = parentStack[i];
+
+            // no need to check checked items
+            if (parent.hasIndentedCodeblock) {
+              break;
+            }
+
+            if (parent.type === "list") {
+              parent.hasIndentedCodeblock = true;
+            }
+          }
+        }
+      }
+      return node;
+    });
+}
+
+function markAlignedList(originalText, options) {
+  return () => ast =>
+    map(ast, (node, index, parentStack) => {
+      if (node.type === "list" && node.children.length !== 0) {
+        // if one of its parents is not aligned, it's not possible to be aligned in sub-lists
+        for (let i = 0; i < parentStack.length; i++) {
+          const parent = parentStack[i];
+          if (parent.type === "list" && !parent.isAligned) {
+            node.isAligned = false;
+            return node;
+          }
+        }
+
+        node.isAligned = isAligned(node);
+      }
+
+      return node;
+    });
+
+  function getListItemStart(listItem) {
+    return listItem.children.length === 0
+      ? -1
+      : listItem.children[0].position.start.column - 1;
+  }
+
+  function isAligned(list) {
+    if (!list.ordered) {
+      /**
+       * - 123
+       * - 123
+       */
+      return true;
+    }
+
+    const [firstItem, secondItem] = list.children;
+
+    const firstInfo = getOrderedListItemInfo(firstItem, originalText);
+
+    if (firstInfo.leadingSpaces.length > 1) {
+      /**
+       * 1.   123
+       *
+       * 1.   123
+       * 1. 123
+       */
+      return true;
+    }
+
+    const firstStart = getListItemStart(firstItem);
+
+    if (firstStart === -1) {
+      /**
+       * 1.
+       *
+       * 1.
+       * 1.
+       */
+      return false;
+    }
+
+    if (list.children.length === 1) {
+      /**
+       * aligned:
+       *
+       * 11. 123
+       *
+       * not aligned:
+       *
+       * 1. 123
+       */
+      return firstStart % options.tabWidth === 0;
+    }
+
+    const secondStart = getListItemStart(secondItem);
+
+    if (firstStart !== secondStart) {
+      /**
+       * 11. 123
+       * 1. 123
+       *
+       * 1. 123
+       * 11. 123
+       */
+      return false;
+    }
+
+    if (firstStart % options.tabWidth === 0) {
+      /**
+       * 11. 123
+       * 12. 123
+       */
+      return true;
+    }
+
+    /**
+     * aligned:
+     *
+     * 11. 123
+     * 1.  123
+     *
+     * not aligned:
+     *
+     * 1. 123
+     * 2. 123
+     */
+    const secondInfo = getOrderedListItemInfo(secondItem, originalText);
+    return secondInfo.leadingSpaces.length > 1;
+  }
 }
 
 const parser = {
