@@ -1,5 +1,7 @@
 "use strict";
 
+const { hasPragma } = require("./pragma");
+
 /*!
  * Extracted from vue codebase
  * https://github.com/vuejs/vue/blob/cfd73c2386623341fdbb3ac636c4baf84ea89c2c/src/compiler/parser/html-parser.js
@@ -16,7 +18,8 @@
 function makeMap(str, expectsLowerCase) {
   const map = Object.create(null);
   const list = str.split(",");
-  for (let i = 0; i < list.length; i++) {
+  const listLength = list.length;
+  for (let i = 0; i < listLength; i++) {
     map[list[i]] = true;
   }
   return expectsLowerCase ? val => map[val.toLowerCase()] : val => map[val];
@@ -147,6 +150,7 @@ function parseHTML(html, options) {
       let text;
       let rest;
       let next;
+
       if (textEnd >= 0) {
         rest = html.slice(textEnd);
         while (
@@ -253,8 +257,7 @@ function parseHTML(html, options) {
   }
 
   function handleStartTag(match) {
-    const tagName = match.tagName;
-    const unarySlash = match.unarySlash;
+    const { tagName, unarySlash } = match;
 
     if (expectHTML) {
       if (lastTag === "p" && isNonPhrasingTag(tagName)) {
@@ -413,7 +416,10 @@ module.exports = {
   parsers: {
     vue: {
       parse,
-      astFormat: "vue"
+      hasPragma,
+      astFormat: "vue",
+      locStart: node => node.start,
+      locEnd: node => node.end
     }
   }
 };
