@@ -26,9 +26,7 @@ function getSortedChildNodes(node, options, resultArray) {
   if (!node) {
     return;
   }
-  const printer = options.printer;
-  const locStart = options.locStart;
-  const locEnd = options.locEnd;
+  const { printer, locStart, locEnd } = options;
 
   if (resultArray) {
     if (node && printer.canAttachComment && printer.canAttachComment(node)) {
@@ -88,8 +86,8 @@ function getSortedChildNodes(node, options, resultArray) {
 // .precedingNode, .enclosingNode, and/or .followingNode properties, at
 // least one of which is guaranteed to be defined.
 function decorateComment(node, comment, options) {
-  const locStart = options.locStart;
-  const locEnd = options.locEnd;
+  const { locStart, locEnd } = options;
+
   const childNodes = getSortedChildNodes(node, options);
   let precedingNode;
   let followingNode;
@@ -179,8 +177,7 @@ function attach(comments, ast, text, options) {
   }
 
   const tiesToBreak = [];
-  const locStart = options.locStart;
-  const locEnd = options.locEnd;
+  const { locStart, locEnd } = options;
 
   comments.forEach((comment, i) => {
     if (
@@ -192,10 +189,7 @@ function attach(comments, ast, text, options) {
     }
 
     decorateComment(ast, comment, options);
-
-    const precedingNode = comment.precedingNode;
-    const enclosingNode = comment.enclosingNode;
-    const followingNode = comment.followingNode;
+    const { precedingNode, enclosingNode, followingNode } = comment;
 
     const pluginHandleOwnLineComment =
       options.printer.handleComments && options.printer.handleComments.ownLine
@@ -299,9 +293,8 @@ function breakTies(tiesToBreak, text, options) {
   if (tieCount === 0) {
     return;
   }
+  const { precedingNode, followingNode } = tiesToBreak[0];
 
-  const precedingNode = tiesToBreak[0].precedingNode;
-  const followingNode = tiesToBreak[0].followingNode;
   let gapEndPos = options.locStart(followingNode);
 
   // Iterate backwards through tiesToBreak, examining the gaps
@@ -499,8 +492,7 @@ function printComments(path, print, options, needsSemi) {
 
   path.each(commentPath => {
     const comment = commentPath.getValue();
-    const leading = comment.leading;
-    const trailing = comment.trailing;
+    const { leading, trailing } = comment;
 
     if (leading) {
       const contents = printLeadingComment(commentPath, print, options);
