@@ -17,7 +17,9 @@ function getSupportInfo(version, opts) {
   );
 
   if (!version) {
-    version = currentVersion;
+    // pre-release version is smaller than the normal version in semver,
+    // we need to treat it as the normal one so as to test new features.
+    version = currentVersion.split("-", 1)[0];
   }
 
   const plugins = opts.plugins;
@@ -95,7 +97,10 @@ function getSupportInfo(version, opts) {
         });
       }
 
-      if (usePostCssParser && language.group === "CSS") {
+      if (
+        usePostCssParser &&
+        (language.name === "CSS" || language.group === "CSS")
+      ) {
         return Object.assign({}, language, {
           parsers: ["postcss"]
         });
