@@ -13,7 +13,8 @@ const internalPlugins = [
   require("./src/language-css"),
   require("./src/language-graphql"),
   require("./src/language-markdown"),
-  require("./src/language-vue")
+  require("./src/language-vue"),
+  require("./src/language-yaml")
 ];
 
 const isArray =
